fix: Ajout de tags qui n'étaient pas gérés Test de génération du contenu des pages
