+++ src/views/pages/session.js
@@ -2,6 +2,7 @@ import { Actionable, Flex, Flexible, Text, View, Wrap } from "@lenra/components"
 import { sessions, speakers } from "../../camping-data.js";
 import { views } from "../../index.gen.js";
 import { days } from "./agenda.js";
+import { buildContentChildren } from "../../utils/contentDescriber.js";
 
 export default function (_data, /* _props,  */{ context: { pathParams } }) {
     const session = sessions[pathParams.key];
@@ -55,75 +56,4 @@ function speakerList(session) {
     if (cards.length > 1)
         return [Wrap(cards).spacing(16)];
     return cards;
-}
-
-/**
- * @param {ContentChild[]} children 
- * @returns
- */
-function buildContentChildren(children) {
-    return children.map(child => {
-        if (typeof child === "string") {
-            if (child === '\n') return null;
-            return Text(child);
-        }
-        switch (child.tag) {
-            case "h1":
-                return Text("")
-                    .children(buildContentChildren(child.children))
-                    .style({
-                        fontSize: 24,
-                        fontWeight: "bold",
-                    });
-            case "h2":
-                return Text("")
-                    .children(buildContentChildren(child.children))
-                    .style({
-                        fontSize: 20,
-                        fontWeight: "bold",
-                    });
-            case "h3":
-                return Text("")
-                    .children(buildContentChildren(child.children))
-                    .style({
-                        fontSize: 18,
-                        fontWeight: "bold",
-                    });
-            case "h4":
-                return Text("")
-                    .children(buildContentChildren(child.children))
-                    .style({
-                        fontSize: 16,
-                        fontWeight: "bold",
-                    });
-            case "p":
-                return Text("")
-                    .children(buildContentChildren(child.children));
-            case "span":
-                if ("children" in child)
-                    return Text("")
-                        .children(buildContentChildren(child.children));
-                return null;
-            case "ul":
-                return Flex(buildContentChildren(child.children))
-                    .direction("vertical")
-                    .padding({ left: 8 });
-            case "li":
-                return Flex([
-                    Text("•"),
-                    Flexible(Flex(buildContentChildren(child.children)))
-                ])
-                    .spacing(8);
-            case "a":
-                console.warn("Link are not supported yet", child)
-                return null;
-                // return Actionable(Text("")
-                //     .children(buildContentChildren(child.children)))
-                //     .onPressed("@lenra:navTo", { path: child.props.href });
-            default:
-                console.warn("Unknown tag", child.tag);
-                return Wrap(buildContentChildren(child.children));
-        }
-    })
-        .filter(child => child !== null);
 }
