--- test/views/pages/session.test.js
@@ -1,9 +1,0 @@
-describe("buildContentChildren", () => {
-    test("check for all sessions", () => {
-        const sessions = require("../../../src/camping-data").sessions;
-        for (const session of Object.values(sessions)) {
-            const children = buildContentChildren(session.children);
-            expect(children).not.toBeEmpty();
-        }
-    });
-});
